PreCourse_1 Exercise_4 - Renamed to GFG.java
